moved connection to state up one level; from HostProfile to index
