+++ src/Components/HostProfileView/HostProfile.js
@@ -7,7 +7,7 @@ class HostProfile extends Component {
   render() {
     const host = this.props.host;
     const hostProfilePic = host.photos[0];
-    const shelter = host.shelterId;
+    const shelter = this.props.shelter
     const shelterProfilePic = shelter.photos[0];
     return (
       <Card>
@@ -35,18 +35,20 @@ class HostProfile extends Component {
   }
 }
 
-const mapStateToProps = (state, props) => {
-  const host = state.hosts.find(host => host._id === props.hostId);
-  const canEdit = state.auth.user._id === host._id
-  return {
-    host: host,
-    canEdit: canEdit
-  };
-};
+// const mapStateToProps = (state, props) => {
+//   const host = state.hosts.find(host => host._id === props.hostId);
+//   const canEdit = state.auth.user._id === host._id
+//   return {
+//     host: host,
+//     canEdit: canEdit
+//   };
+// };
 
-const mapDispatchToProps = null;
+// const mapDispatchToProps = null;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(HostProfile);
+// export default connect(
+//   mapStateToProps,
+//   mapDispatchToProps
+// )(HostProfile);
+
+export default HostProfile
