+++ src/Components/index.js
@@ -1,7 +1,7 @@
 import AnimalListView from "./AnimalListView";
 import { AnimalProfileView } from "./AnimalProfileView";
 import HostListView from "./HostListView";
-import { HostProfileView } from "./HostProfileView";
+import HostProfileView from "./HostProfileView";
 import Login from "./LoginAndRegistration/Login";
 import { Nav } from './Nav';
 import { PublicView } from "./PublicView";
